refactor: extract routers/api.py
Move /contexts/{ctx}/question, /contexts/{ctx}/evaluate,
/contexts/{ctx}/questions, /api/questions/{ctx}, /api/attempts, /health
out of main.py into api/routers/api.py. Wire router into main.py.
